+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- networks: Ignore the PCI `slot_path` of wireless devices from data provided by
+  TurrisHW. This is a stopgap measure until issue #252 will be resolved.
+
 ## [5.5.2] - 2023-05-09
 ### Fixed
 - lan: Unify DHCPv4/6 lease timestamps to the "end of the lease" format instead